NetTruder - GA + FCM + RF
Implements genetic algorithm on the dataset to extract relevant features, then FCM is performed on the KDD-98 dataset. Finally the outputs of GA and Fuzzy C Means Clustering algorithm is clubbed and passed on to a deep learning algorthm, in this case Random Forest Classifier. The result yielded a 99.9% accuracy.